databene-commons-core/src/main/java/org/databene/commons/converter/NumberFormatter.java
/*
 * (c) Copyright 2010-2012 by Volker Bergmann. All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without
 * modification, is permitted under the terms of the
 * GNU General Public License (GPL).
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
 * WITHOUT A WARRANTY OF ANY KIND. ALL EXPRESS OR IMPLIED CONDITIONS,
 * REPRESENTATIONS AND WARRANTIES, INCLUDING ANY IMPLIED WARRANTY OF
 * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE OR NON-INFRINGEMENT, ARE
 * HEREBY EXCLUDED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
 * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
 * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
 * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
 * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
 * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
 * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 * POSSIBILITY OF SUCH DAMAGE.
 */

package org.databene.commons.converter;

import java.text.NumberFormat;

import org.databene.commons.ConversionException;

/**
 * Formats {@link Number} objects as {@link String} using {@link NumberFormat}.<br/><br/>
 * Created: 26.02.2010 08:36:36
 * @since 0.5.0
 * @author Volker Bergmann
 */
public class NumberFormatter extends NumberFormatBasedConverter<Number, String> {
	
    // constructors ----------------------------------------------------------------------------------------------------

    public NumberFormatter() {
		super(Number.class, String.class);
	}

	public NumberFormatter(String pattern) {
		super(Number.class, String.class, pattern);
	}
	
	// Converter interface implementation ------------------------------------------------------------------------------

	@Override
	public String convert(Number value) throws ConversionException {
		return format(value);
	}
	
}
